refactor: move validator drain to validator_pool.rs
